refs #2307, #2318: Introduce T_operational parameter
T_ref is used both as reference temperature (for the linear temperature dependency of the resistance/conductance) and as operational temperature (if no conditional heat port is available). With the introduction of T_operational the confusion can be avoided by a clean separation of the two temperatures.

Furthermore, parameter T can now be modified as final.