Perform syntax check using travis-ci
This uses OpenModelica v1.12.0 and Maplesoft's moparse to perform syntax
checking of the library. The OpenModelica job returns error on warnings
when it detects mismatching uses-annotations (which is not foolproof),
bad package.order, etc. The job could be enhanced to perform additional
checks or calculate some statistics such as number of classes in the
future.