Fixing Modelica.Math.Matrices.Utilities.toUpperHessenberg.
This addresses a couple of issues with toUpperHessenberg.
First, in the previous implementation, the elements below the subdiagonal of H was never initialized. This is an error by the Modelica Specification 3.4, section 12.4.4.
Second, there is no point in copying different parts of H from A or Aout, it just makes for confusing code. Even if ilo and ihi are set to nontrivial values, the upper Hessenberg matrix will be in Aout's upper triangle and first subdiagonal.