Set Evaluate=true for useQuaternions
There are if-equations that would be unbalanced unless useQuaternions is evaluated during translation; using Evaluate=true makes this evaluation somewhat more transparent to users, and tools don't have to go all the way to determining equation sizes in order to see that useQuaternions must be evaluated.

This is the fallback solution that is applied because turning useQuaternions into 'constant' was rejected in https://github.com/modelica/ModelicaStandardLibrary/pull/3113.