Set Evaluate=true for enforceStates
State selection happens during translation, so it makes no sense to pretend that these are normal parameters that might not require evaluation during translation.

This is the fallback solution that is applied because turning enforceStates into 'constant' was rejected in https://github.com/modelica/ModelicaStandardLibrary/pull/3113.

There was one case of enforceStates with Evaluate=true before this change, so this also fixes a consistency issue.