Set Evaluate=true for 'exact'
This is the fallback solution that is applied because turning 'exact' into 'constant' was rejected in https://github.com/modelica/ModelicaStandardLibrary/pull/3113.

As noted in the affected MultiBody parts, the 'exact' variable has impact on the structure of the equations.  In order to get the desired performance benefit, the tool needs to know that it is allowed and expected to use the value of 'exact' during translation.

For the Rotational and Translational parts, there are initial if-equations that become unbalanced unless 'exact' is evaluated during translation.  This is made more transparent to the user by setting the Evaluate=true just like in MultiBody.