Improve conversion
In 49460b02f3cea449133eda570e04932ae1a3a08a for #2898 the default initType of PID was converted from InitPID.DoNotUse_InitialIntegratorState to Init.InitialState. For LimPID it was different, that is from InitPID.DoNotUse_InitialIntegratorState to Init.NoInit. The conversion rule was added to convert from InitPID.DoNotUse_InitialIntegratorState to Init.NoInit, matching the LimPID manual conversion.

It was reported that the conversion to Init.NoInit breaks the initialization and model structure of user models. Despite Init.NoInit being the default initType for PI, the default initialization for PID and LimPID should better be converted to Init.InitialState. No matter if Init.NoInit or Init.InitialState, the initialization of the converted PID or LimPID component behaves differently since there is no compatibility preserving conversion possible.