Avoid singularity close to the solution
Avoid singularity close to the solution by not having diagonal(R_T)=ones(3),
which is basically x^2 and thus its derivative is singular at the solution.
Instead use off-diagonal(R_T)=zeros(3) whose derivative is non-singular at the solution.
(With some trick so that we only get {1,1,1} on the diagonal and not {-1,-1,1}.)