Remove Evaluate=true for fixed=false parameter 'n_b'
From the brief discussion in https://github.com/modelica/ModelicaSpecification/issues/2951, it seems that `Evaluate = true` on a parameter with `fixed = false` is not expected to have any effect.  Thus I'd like to propose that the `Evaluate = true` is removed in order to get better correspondence between presence of `Evaluate = true` and parameters actually becoming evaluated.